AG-6815 - Push Axis domain calculations into Axis class hierarchy.
