+++ charts-packages/ag-charts-community/src/chart/axis/numberAxis.ts
@@ -114,4 +114,14 @@ export class NumberAxis extends ChartAxis {
             return String(datum);
         }
     }
+
+    protected updateDomain(domain: any[], isYAxis: boolean, primaryTickCount?: number) {
+        if (isYAxis) {
+            // the `primaryTickCount` is used to align the secondary axis tick count with the primary
+            this.setDomain(domain, primaryTickCount);
+            return primaryTickCount ?? this.scale.ticks!(this.calculatedTickCount).length;
+        }
+
+        return super.updateDomain(domain, isYAxis, primaryTickCount);
+    }
 }
